pyproject.toml
[project]
name = "hitl-adk"
version = "0.1.0"
description = "Human-in-the-Loop ADK Agent with VertexAI Session and Memory Services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[vertexai]>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hitl_agent"]

[tool.adk]
agent = "hitl_agent:root_agent"

